+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Fixed
+* Typos in clipboard checks added in 1.4.2 (require != requires)
 
 
 ## [1.4.2] - 2022-02-15